+++ app.js
@@ -1,17 +1,16 @@
 const express = require('express');
-const cors = require('cors');
 const request = require('request');
 
 let app = express();
 app.use(express.static(__dirname + '/public'))
-    .use(cors())
     .use(express.urlencoded({ extended: true }))
     .use(express.json());
     
 
-require('./Components/htmlRoutes')(app);
-require('./Components/auth')(app);
-require('./Components/get_playlists')(app);
+require('./routes/htmlRoutes')(app);
+require('./routes/api_routes/auth')(app);
+require('./routes/api_routes/add_playlists')(app);
+require('./routes/api_routes/device')(app);
 
 console.log('Listening on 8888');
 app.listen(8888);
